Add an "ink_story" macro for running Inklecate
NOTE: I ran into a problem in Bazel when implementing this.
This macro is intended to be used from an external workspace,
i.e. somebody's Ink game project. But the macro is expanded
in the context of that external workspace. That makes it tricky
to refer unambiguously to the "inklecate" tool, which lives in
*this* workspace. "@ink//:inklecate" works iff we are imported
as "@ink", but we have no way to enforce that.